requirements.txt
Django==1.7a2
django-pipeline==1.3.22
cssmin==0.2.0
jsmin==2.0.9
Pillow==2.3.0
psycopg2==2.5.2
pytz==2013.9
